Enum addr2line::LookupResult

pub enum LookupResult<L: LookupContinuation> {
    Load {
        load: SplitDwarfLoad<<L as LookupContinuation>::Buf>,
        continuation: L,
    },
    Output(<L as LookupContinuation>::Output),
}

Operations that consult debug information may require additional files to be loaded if split DWARF is being used. This enum returns the result of the operation in the Break variant, or information about the split DWARF that is required and a continuation to invoke once it is available in the Continue variant.

This enum is intended to be used in a loop like so:

  const ADDRESS: u64 = 0xdeadbeef;
  let mut r = ctx.find_frames(ADDRESS);
  let result = loop {
    match r {
      LookupResult::Output(result) => break result,
      LookupResult::Load { load, continuation } => {
        let dwo = do_split_dwarf_load(load);
        r = continuation.resume(dwo);
      }
    }
  };

Variants

Load

Fields

load: SplitDwarfLoad<<L as LookupContinuation>::Buf>

The information needed to find the split DWARF data.

continuation: L

The continuation to resume with the loaded split DWARF data.

The lookup requires split DWARF data to be loaded.

Output(<L as LookupContinuation>::Output)

The lookup has completed and produced an output.

Implementations

impl<L: LookupContinuation> LookupResult<L>

pub fn skip_all_loads(self) -> L::Output

Callers that do not handle split DWARF can call skip_all_loads to fast-forward to the end result. This result is produced with the data that is available and may be less accurate than the the results that would be produced if the caller did properly support split DWARF.