addr2line

Function demangle_auto

pub fn demangle_auto(
    name: Cow<'_, str>,
    language: Option<DwLang>,
) -> Cow<'_, str>

Apply ‘best effort’ demangling of a symbol name.

If language is given, then only the demangling scheme for that language is used.

If language is None, then heuristics are used to determine how to demangle the name. Currently, these heuristics are very basic.

If demangling fails or is not required, then name is returned unchanged.