Module addr2line::lookup

Structs

LoopingLookup 🔒

Some functions (e.g. find_frames) require considering multiple compilation units, each of which might require their own split DWARF lookup (and thus produce a continuation).

MappedLookup 🔒

SimpleLookup 🔒

SplitDwarfLoad

This struct contains the information needed to find split DWARF data and to produce a gimli::Dwarf<R> for it.

Enums

LookupResult

Operations that consult debug information may require additional files to be loaded if split DWARF is being used. This enum returns the result of the operation in the Output variant, or information about the split DWARF that is required and a continuation to invoke once it is available in the Load variant.

Traits

LookupContinuation

This trait represents a partially complete operation that can be resumed once a load of needed split DWARF data is completed or abandoned by the API consumer.