Trait addr2line::LookupContinuation

pub trait LookupContinuation: Sized {
    type Output;
    type Buf: Reader;

    // Required method
    fn resume(self, input: Option<Arc<Dwarf<Self::Buf>>>) -> LookupResult<Self>;
}

This trait represents a partially complete operation that can be resumed once a load of needed split DWARF data is completed or abandoned by the API consumer.

Required Associated Types

type Output

The final output of this operation.

type Buf: Reader

The type of reader used.

Required Methods

fn resume(self, input: Option<Arc<Dwarf<Self::Buf>>>) -> LookupResult<Self>

Resumes the operation with the provided data.

After the caller loads the split DWARF data required, call this method to resume the operation. The return value of this method indicates if the computation has completed or if further data is required.

If the additional data cannot be located, or the caller does not support split DWARF, resume(None) can be used to continue the operation with the data that is available.

Object Safety

This trait is not object safe.